Render: split the cube geometry out of the type checker
The SVG rendering of a cube has a sizeable part that knows nothing about
terms: the projection matrices, the camera, and the drawing of the cube
itself, which needs only a function saying what to put on each vertex,
edge and face. That part sat in the middle of TypeCheck.hs, between the
tope machinery and the LSP hover types.

Move it to Rzk.Render.Geometry. Nothing outside TypeCheck.hs referred to
it, so this is a pure move; the numeric defaulting the old file silenced
wholesale is now written out at the four places that relied on it.

The module mentions no variable representation, so both cores can share
it while the free-foil migration is in flight.
